the rst uses tabs to link to tests corresponding to the example shown
