+++ docs/examples/request_data/request_data_7.py
@@ -6,7 +6,6 @@ from litestar import Litestar, MediaType, post
 from litestar.datastructures import UploadFile
 from litestar.enums import RequestEncodingType
 from litestar.params import Body
-from litestar.testing import TestClient
 
 
 @post(path="/", media_type=MediaType.TEXT, sync_to_thread=False)
@@ -19,10 +18,3 @@ def handle_file_upload(
 
 
 app = Litestar(route_handlers=[handle_file_upload])
-
-
-def test_file_upload() -> None:
-    with TestClient(app) as client:
-        response = client.post("/", files={"file": ("hello.txt", b"hello")})
-        assert response.status_code == 201
-        assert response.text == "hello.txt, " + sha256(b"hello").hexdigest()
