+++ docs/examples/request_data/request_data_9.py
@@ -7,7 +7,6 @@ from litestar import Litestar, post
 from litestar.datastructures import UploadFile
 from litestar.enums import RequestEncodingType
 from litestar.params import Body
-from litestar.testing import TestClient
 
 
 @post(path="/")
@@ -23,10 +22,3 @@ async def handle_file_upload(
 
 
 app = Litestar(route_handlers=[handle_file_upload])
-
-
-def test_file_upload() -> None:
-    with TestClient(app) as client:
-        response = client.post("/", files={"file": ("hello.txt", b"hello")})
-        assert response.status_code == 201
-        assert response.json().get("hello.txt") == sha256(b"hello").hexdigest()
